make options.expires take milliseconds not days
Using milliseconds gives the developer the granularity needed without
requiring creation of a Date() object, and it is easy to calculate
larger time units from milliseconds when desired.  It is also the same
time unit used in standard cookies, so it maintains the paradigm.  Can
still take a standard Date() object as well.